Move GoogleLoginButton component to /shared directory
