+++ src/client/app/components/shared/GoogleLoginButton/GoogleLoginButton.jsx
